requirements.txt
cassandra-driver==3.17
matplotlib==3.0.2
